Created abstract class User
