+++ src/Model/User.java
@@ -5,24 +5,25 @@
  */
 package Model;
 
+import java.io.InputStream;
+
 /**
  *
  * @author Samuel
  */
-public class User {
+public abstract  class User {
     
-    private Long id;
-    private String name;
-    private String login;
-    private String password;
-    private ImageFile perfilImage;
-    private String email;
-    private String address;
-    private String CEP;
-    private String phone;
-    private String school;
+    protected Long id;
+    protected String name;
+    protected String login;
+    protected String password;
+    protected ImageFile perfilImage;
+    protected String email;
+    protected String address;
+    protected String CEP;
+    protected String phone;
 
-    public User(Long id, String name, String login, String password, ImageFile perfilImage, String email, String address, String CEP, String phone, String school) {
+    public User(Long id, String name, String login, String password, ImageFile perfilImage, String email, String address, String CEP, String phone) {
         this.id = id;
         this.name = name;
         this.login = login;
@@ -32,10 +33,9 @@ public class User {
         this.address = address;
         this.CEP = CEP;
         this.phone = phone;
-        this.school = school;
     }
 
-    public User(String name, String login, String password, ImageFile perfilImage, String email, String address, String CEP, String phone, String school) {
+    public User(String name, String login, String password, ImageFile perfilImage, String email, String address, String CEP, String phone) {
         this.name = name;
         this.login = login;
         this.password = password;
@@ -44,13 +44,92 @@ public class User {
         this.address = address;
         this.CEP = CEP;
         this.phone = phone;
-        this.school = school;
     }
 
     public User() {
         perfilImage = new ImageFile("");
     }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
     
+    public void setId(int id) {
+        this.id = new Long(id);
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getLogin() {
+        return login;
+    }
+
+    public void setLogin(String login) {
+        this.login = login;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public ImageFile getPerfilImage() {
+        return perfilImage;
+    }
+
+    public void setPerfilImage(ImageFile perfilImage) {
+        this.perfilImage = perfilImage;
+    }
     
-    
+        public void setPerfilImage(InputStream inputS, String name) {
+        
+        this.perfilImage = new ImageFile(inputS, name);
+        
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public String getCEP() {
+        return CEP;
+    }
+
+    public void setCEP(String CEP) {
+        this.CEP = CEP;
+    }
+
+    public String getPhone() {
+        return phone;
+    }
+
+    public void setPhone(String phone) {
+        this.phone = phone;
+    }
+
 }
